Refactor to use new assessment renderer API. Clean up interactions with perseus itemRenderer.
