+++ kolibri_exercise_perseus_plugin/package.json
@@ -13,7 +13,6 @@
     "react-addons-css-transition-group": "0.14.7",
     "react-addons-pure-render-mixin": "0.14.7",
     "react-dom": "0.14.7",
-    "seededshuffle": "0.1.1",
     "string-replace-loader": "^1.0.5",
     "underscore": "1.4.4"
   }
